Configuration for any form.
Default to comments, registration and contact form.
But any form can be configured.

Removed the tabs config page since it is such a small form there is no need for it.

Preprocess every pattern to leave it ready for matching. Separate exact from regex matching to make to more efficient.